Pivot from the poorly named "internal" handlers. - Keep the cripple->disabled renaming but move the handler back to config. - Move the new GraphCacheMetrics stuff as the first of possibly more dedicated   "/test" endpoints.
