+++ handlers/testing.go
@@ -9,6 +9,8 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/prometheus/client_golang/prometheus"
+	dto "github.com/prometheus/client_model/go"
 	core_v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -17,6 +19,7 @@ import (
 	"github.com/kiali/kiali/config"
 	"github.com/kiali/kiali/handlers/authentication"
 	"github.com/kiali/kiali/kubernetes"
+	"github.com/kiali/kiali/prometheus/internalmetrics"
 	"github.com/kiali/kiali/util"
 )
 
@@ -50,4 +53,36 @@ func WithAuthInfo(authInfo map[string]*api.AuthInfo, hf http.HandlerFunc) http.H
 func WithFakeAuthInfo(conf *config.Config, hf http.HandlerFunc) http.HandlerFunc {
 	authInfo := map[string]*api.AuthInfo{conf.KubernetesConfig.ClusterName: {Token: "test"}}
 	return WithAuthInfo(authInfo, hf)
+}
+
+// GraphCacheMetrics holds graph cache metrics that can be exposed via the API.
+// This is useful for testing and monitoring without having to scrape the Prometheus metrics endpoint.
+type GraphCacheMetrics struct {
+	GraphCacheEvictions float64 `json:"graphCacheEvictions"`
+	GraphCacheHits      float64 `json:"graphCacheHits"`
+	GraphCacheMisses    float64 `json:"graphCacheMisses"`
+}
+
+// getCounterValue extracts the current value from a Prometheus Counter.
+func getCounterValue(counter prometheus.Counter) float64 {
+	m := &dto.Metric{}
+	if err := counter.Write(m); err != nil {
+		return 0
+	}
+	return m.Counter.GetValue()
+}
+
+// GraphCacheMetricsHandler returns Kiali's graph cache metrics in JSON format.
+// This endpoint provides a simple way to access internal metrics without having to
+// parse the Prometheus text format from the /metrics endpoint.
+func GraphCacheMetricsHandler() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		metrics := GraphCacheMetrics{
+			GraphCacheEvictions: getCounterValue(internalmetrics.GetGraphCacheEvictionsTotalMetric()),
+			GraphCacheHits:      getCounterValue(internalmetrics.GetGraphCacheHitsTotalMetric()),
+			GraphCacheMisses:    getCounterValue(internalmetrics.GetGraphCacheMissesTotalMetric()),
+		}
+
+		RespondWithJSON(w, http.StatusOK, metrics)
+	}
 }
